move functions in classes to separate test files
